--- front-end-mobile/src/features/authentication/forgotPassword/index.ts
@@ -1,1 +1,0 @@
-export { default as ForgotPasswordScreen } from './ForgotPasswordScreen';
