--- front-end-mobile/src/features/authentication/signIn/index.ts
@@ -1,1 +1,0 @@
-export { default as SignInScreen } from './SignInScreen';
