--- front-end-mobile/src/features/authentication/signUp/index.ts
@@ -1,1 +1,0 @@
-export { default as SignUpScreen } from './SignUpScreen';
